CS7NS1-Sclable Computing/Practical-4/scalable-ass-5/requirements.txt
git+git://github.com/blockstack/secret-sharing.git@master#egg=secretsharing
pycrypto
